[tooling] remove Node from our tooling and only use Deno for JS & TS
