+++ .github/workflows/main.yml
@@ -22,7 +22,7 @@ jobs:
     - uses: actions/checkout@v1
 
     # Install missing language compilers and interpreters
-    # The default image natively supports bash, c/c++, go, java, julia, node, python, ruby, rust
+    # The default image natively supports bash, c/c++, go, java, julia, deno, python, ruby, rust
     # For new languages, see https://github.com/actions/virtual-environments/tree/d18c405/images/linux first
 
     - uses: denolib/setup-deno@v2
@@ -41,19 +41,6 @@ jobs:
         unzip -d /tmp /tmp/v_linux.zip
         sudo /tmp/v/v symlink
         # We can not remove /tmp/v/ as /usr/local/bin/v needs some files in there
-
-    # Install missing dependencies
-
-    # Install npm packages if not cached
-    - name: cache-npm
-      uses: actions/cache@v1
-      id: cache-npm
-      with:
-        path: node_modules
-        key: ${{ runner.os }}-${{ hashFiles('package-lock.json') }}
-    - name: deps-npm
-      if: steps.cache-npm.outputs.cache-hit != 'true'
-      run: npm install
 
     # Install python packages if not cached
     - name: cache-python
