+++ tool/templates/template.js
@@ -1,5 +1,3 @@
-const { performance } = require("perf_hooks");
-
 /**
  * @param {string} s puzzle input in string format
  * @returns {string} solution flag
@@ -8,8 +6,8 @@ const run = (s) => {
   // Your code goes here
 };
 
-const start = performance.now();
-const answer = run(process.argv[2]);
+let start = performance.now();
+let answer = run(Deno.args[0]);
 
 console.log(`_duration:${performance.now() - start}`);
 console.log(answer);
