--- tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "compilerOptions": {
-    "module": "commonjs",
-    "target": "es2019",
-    "noImplicitAny": true,
-    "allowJs": true,
-    "strict": true,
-    "noEmit": true
-  }
-}
